feat(agent): fix behaviors and implement missing features.
