+++ agent/backend/otel/otel.go
@@ -14,7 +14,6 @@ import (
 	"go.opentelemetry.io/collector/receiver"
 	"go.uber.org/zap"
 	"os"
-	"strings"
 	"time"
 )
 
@@ -89,14 +88,6 @@ func (o openTelemetryBackend) Configure(logger *zap.Logger, repo policies.Policy
 	return nil
 }
 
-func (o openTelemetryBackend) SetCommsClient(agentID string, client *mqtt.Client, baseTopic string) {
-	o.mqttClient = client
-	otelBaseTopic := strings.Replace(baseTopic, "?", "otlp", 1)
-	o.otlpMetricsTopic = fmt.Sprintf("%s/m/%c", otelBaseTopic, agentID[0])
-	o.otlpTracesTopic = fmt.Sprintf("%s/t/%c", otelBaseTopic, agentID[0])
-	o.otlpLogsTopic = fmt.Sprintf("%s/l/%c", otelBaseTopic, agentID[0])
-}
-
 func (o openTelemetryBackend) Version() (string, error) {
 	executable, err := memexec.New(openTelemtryContribBinary)
 	if err != nil {
@@ -143,6 +134,7 @@ func (o openTelemetryBackend) Start(ctx context.Context, cancelFunc context.Canc
 	}
 	o.runningCollectors = make(map[string]context.Context)
 	o.mainCancelFunction = cancelFunc
+	o.mainContext = ctx
 	o.startTime = time.Now()
 
 	currentVersion, err := o.Version()
@@ -169,7 +161,7 @@ func (o openTelemetryBackend) Start(ctx context.Context, cancelFunc context.Canc
 	return nil
 }
 
-func (o openTelemetryBackend) Stop(ctx context.Context) error {
+func (o openTelemetryBackend) Stop(_ context.Context) error {
 	o.logger.Info("stopping all running policies")
 	o.mainCancelFunction()
 	for policyID, policyCtx := range o.runningCollectors {
@@ -179,9 +171,22 @@ func (o openTelemetryBackend) Stop(ctx context.Context) error {
 	return nil
 }
 
-func (o openTelemetryBackend) FullReset(ctx context.Context) error {
-	//TODO implement me
-	panic("implement me")
+func (o openTelemetryBackend) FullReset(_ context.Context) error {
+	o.logger.Info("resetting all policies and restarting")
+	for policyID, policyCtx := range o.runningCollectors {
+		o.logger.Debug("stopping policy context", zap.String("policy_id", policyID))
+		policyCtx.Done()
+		policy, err := o.policyRepo.Get(policyID)
+		if err != nil {
+			o.logger.Error("failed to get policy", zap.Error(err))
+			return err
+		}
+		err = o.ApplyPolicy(policy, true)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
 }
 
 func (o openTelemetryBackend) GetStartTime() time.Time {
@@ -205,97 +210,4 @@ func (o openTelemetryBackend) GetRunningStatus() (backend.RunningStatus, string,
 		return backend.Running, fmt.Sprintf("opentelemetry backend running with %d policies", amountCollectors), nil
 	}
 	return backend.Offline, "opentelemetry backend offline, waiting for policy to come to start running", nil
-}
-
-func (o openTelemetryBackend) ApplyPolicy(newPolicyData policies.PolicyData, updatePolicy bool) error {
-	if !o.policyRepo.Exists(newPolicyData.ID) {
-		temporaryFile, err := os.CreateTemp(o.policyConfigDirectory, fmt.Sprintf("otel-%s-config.yml", newPolicyData.ID))
-		if err != nil {
-			return err
-		}
-		err = o.addRunner(newPolicyData, temporaryFile.Name())
-		if err != nil {
-			return err
-		}
-	} else {
-		currentPolicyData, err := o.policyRepo.Get(newPolicyData.ID)
-		if err != nil {
-			return err
-		}
-		if currentPolicyData.Version <= newPolicyData.Version {
-			dataAsByte := []byte(newPolicyData.Data.(string))
-			currentPolicyPath := o.policyConfigDirectory + fmt.Sprintf("otel-%s-config.yml", currentPolicyData.ID)
-			o.logger.Info("new policy version received, updating", zap.String("policy_id", newPolicyData.ID), zap.Int32("version", newPolicyData.Version))
-			err := os.WriteFile(currentPolicyPath, dataAsByte, os.ModeTemporary)
-			if err != nil {
-				return err
-			}
-			err = o.policyRepo.Update(newPolicyData)
-			if err != nil {
-				return err
-			}
-			o.otelReceiverTaps = append(o.otelReceiverTaps, newPolicyData.ID)
-		} else {
-			o.logger.Info("current policy version is newer than the one being applied, skipping",
-				zap.String("policy_id", newPolicyData.ID),
-				zap.Int32("current_version", currentPolicyData.Version),
-				zap.Int32("incoming_version", newPolicyData.Version))
-		}
-	}
-
-	return nil
-}
-
-func (o openTelemetryBackend) addRunner(policyData policies.PolicyData, policyFilePath string) error {
-	policyContext := context.WithValue(o.mainContext, "policy_id", policyData.ID)
-	executable, err := memexec.New(openTelemtryContribBinary)
-	if err != nil {
-		return err
-	}
-	defer func(executable *memexec.Exec) {
-		err := executable.Close()
-		if err != nil {
-			o.logger.Error("error closing executable", zap.Error(err))
-		}
-	}(executable)
-	command := executable.CommandContext(policyContext, "--config", policyFilePath)
-	stderr, err := command.StderrPipe()
-	if err != nil {
-		return err
-	}
-	go func(ctx context.Context) {
-		scanner := bufio.NewScanner(stderr)
-		for scanner.Scan() {
-			o.logger.Info("stderr output",
-				zap.String("policy_id", policyData.ID),
-				zap.String("line", scanner.Text()))
-			if command.Err != nil {
-				o.logger.Error("error running command", zap.Error(command.Err))
-				ctx.Done()
-				return
-			}
-		}
-	}(policyContext)
-
-	return nil
-}
-
-func (o openTelemetryBackend) addPolicyControl(policyCtx context.Context, policyID string) {
-	o.runningCollectors[policyID] = policyCtx
-}
-
-func (o openTelemetryBackend) removePolicyControl(policyID string) {
-	o.runningCollectors[policyID].Done()
-	delete(o.runningCollectors, policyID)
-}
-
-func (o openTelemetryBackend) RemovePolicy(data policies.PolicyData) error {
-	if o.policyRepo.Exists(data.ID) {
-		o.removePolicyControl(data.ID)
-		err := o.policyRepo.Remove(data.ID)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
